Completed filling the test space (390 cases) and added results for print1D.
Also fixed printing of empty expressions.

Changed the associativity of the exponentiation operator ^ to right, but didn't
update the code so the corresponding test cases won't pass yet.
This was to ease correct printing in 2D while preserving a uniform handling of
precedence and associativity with 1D printing and parsing.

--HG--
branch : testing