Fixed a bug where dimensioning lines disappeared under certain conditions.
Although I had to fix some other minor issues as well, the real problem was that
std::minmax can't be used to sort/swap the values of two variables without
special care. The problem is that it returns a pair of references to the
variables and so the value of the one will get overwritten before it is used to
set the value of the other.

--HG--
branch : bugfix