Removed Visual Studio projects and solution in preparation for porting to CMake.
The plan is to create CMake scripts for the project that can be used to generate
a wide range of build scripts and IDE projects for every platform.

Once Visual Studio 2012 projects and solution can be generated on demand, there
will be no reason to keep tracking them using Mercurial.

Moreover the project file structure will be revised before moving to the new
build system and there is no point dragging those files around.

The downside is that the project will remain without build scripts for a few
revisions.

--HG--
branch : regression