Created a CMake configuration file to build the project.
This will enable to build the project on any system CMake supports using a
platform-independent set of commands. Moreover it will make version requirements
for any dependencies explicit and machine-readable!

## Successful Builds

1. On Linux Mint 18 Sarah (Linux base 4.4.0-21-generic x86_64)

    1. Using clang @ 3.8.0-2ubuntu4 (tags/RELEASE_380/final)
    2. Using gcc @ 5.4.0 20160609 (Ubuntu 5.4.0-6ubuntu1~16.04.9)

   With the following packages installed:

    1. libglew-dev @ 1.13.0-2
    2. freeglut3 @ 2.8.1-2
    3. libeigen3-dev @ 3.3~beta1-2
    4. libboost-all-dev @ 1.58.0.1ubuntu1
    5. cmake @ 3.5.1-1ubuntu3

Currently tests are run as part of the executable initialization sequence and
they pass. They will start running as part of the build soon.

--HG--
branch : build