Replaced GL_POLYGON_SMOOTH with MSAA to avoid a bug in recent NVIDIA's drivers.
There is a bug in recent versions of the NVIDIA drivers causing visual artefacts
when rendering with glRect* and GL_POLYGON_SMOOTH enabled. In particular, the
rectangle is drawn as two separate triangles with a small but visible gap
between them.

In order to work around this bug while keeping visual quality high, I disabled
blend-based antialiasing and enabled multisample antialiasing (MSAA) as a cheep
way to achieve full-screen antialiasing (FSAA).

## Successful Builds

1. On Linux Mint 18 Sarah (Linux base 4.4.0-21-generic x86_64)

    1. Using clang @ 3.8.0-2ubuntu4 (tags/RELEASE_380/final)
    2. Using gcc @ 5.4.0 20160609 (Ubuntu 5.4.0-6ubuntu1~16.04.9)

   With the following packages installed:

    1. libglew-dev @ 1.13.0-2
    2. freeglut3 @ 2.8.1-2
    3. libeigen3-dev @ 3.3~beta1-2
    4. libboost-all-dev @ 1.58.0.1ubuntu1
    5. cmake @ 3.5.1-1ubuntu3

--HG--
branch : maintenance