Changed the character used to render quotient lines in text output to '_'.
This was done to fix a failing test suite.

Originally I developed code and unit tests on a Windows PC which used the IBM PC
character set (Code page 437) in the command-line. There I pasted a literal '─'
(\u2500) in the test output strings, but somehow the UTF-8 representation for
a 'Δ' (\u0394) got written to the file! Moreover, I used the octal value '\304'
in the function to be tested. The correct glyph displayed and tests passed and
everything was good.

Unfortunately, when I changed system, the new compiler interpreted both values
in different ways causing the tests to fail.

In the long term, all string processing in the code base will be updated to work
with Unicode, but this might be delayed until a proper font engine is
adopted.

For now, I'm dropping support for anything above ASCII and using underscores to
draw the quotient lines.

## Successful Builds

1. On Linux Mint 18 Sarah (Linux base 4.4.0-21-generic x86_64)

    1. Using clang @ 3.8.0-2ubuntu4 (tags/RELEASE_380/final)
    2. Using gcc @ 5.4.0 20160609 (Ubuntu 5.4.0-6ubuntu1~16.04.9)

   With the following packages installed:

    1. libglew-dev @ 1.13.0-2
    2. freeglut3 @ 2.8.1-2
    3. libeigen3-dev @ 3.3~beta1-2
    4. libboost-all-dev @ 1.58.0.1ubuntu1
    5. cmake @ 3.5.1-1ubuntu3

Currently tests are run as part of the executable initialization sequence and
they pass. They will start running as part of the build soon.

--HG--
branch : portability