Added a distclean target to CMakeLists.txt.
The corresponding command also executes a standard clean and will ease removal
of binary and test directories and purging any configuration for the generated
application's CMake project.

## Testing Done

Built the new target invoking cmake from inside and outside the build tree and
checked that the correct files are removed. Then built the project and tests
again to ensure there are not errors.

Repeated the tests on:

1. Linux Mint 18 Sarah 64-bit physical machine.
    1. Using makefiles and gcc 5.4.0.
    2. Using makefiles and clang 3.8.0.
2. Windows 7 64-bit virtual machine.
    1. Using Visual Studio 2017 Community Edition.
3. Windows 8.1 64-bit physical machine.
    1. Using Visual Studio 2017 Community Edition.

No problems observed.

--HG--
branch : build