Added version tags starting with v0.0.0 and increasing for features and bugfixes
The idea is to comply with [SemVer 2.0.0](https://semver.org/), but I have no
stable public API yet, so major version gets 0 and the standard no longer
prescribes a specific way to increase minor and patch versions. So I increase
minor version when significant new functionality is added and patch version
when bugs are fixed.

The situation is slightly complicated by the fact that this repository currently
hosts both LostArt (an application) and Graphene (a library) causing
ambiguities. e.g. "Should this commit be considered a refactoring of the
application or an enhancement of the library?"

There was a single tag from an older but similar versioning scheme that was
removed to avoid confusion. It was never published, so removing it shouldn't
cause any problems.

Also added a tag at what most probably has been the version presented at softlab
in September 2014, after initial development.

--HG--
branch : administrative