Added a Travis CI configuration file to the project.
This will enable continuous integration for the GitHub mirror.

Had to build boost manually since the most recent version available in Travis CI
whitelist was 1.55. (Actually a few more recent packages do appear in the
whitelist, but the [instructions][1] for identifying the APT source did not
work.)

[1]: https://docs.travis-ci.com/user/installing-dependencies/#Identifying-the-source-for-a-missing-package

Considered caching the build output, but since only the headers are needed, it's
faster without using the cache. (b2 just copies the headers to an installation
directory. It also builds 'test', but that's just because there was no option to
build nothing.)

Another complication was the need to initialize glut during testing. It was
handled by using xvfb as explained in Travis CI [documentation][2].

[2]: https://docs.travis-ci.com/user/gui-and-headless-browsers/#Using-xvfb-to-Run-Tests-That-Require-a-GUI

--HG--
branch : administrative